V3

+++ src/Controller/TicketStatutController.php
@@ -2,8 +2,10 @@
 
 namespace App\Controller;
 
+use App\Entity\GlobalStatut;
 use App\Entity\TicketStatut;
 use App\Form\TicketStatutType;
+use Doctrine\ORM\QueryBuilder;
 use Omines\DataTablesBundle\Adapter\Doctrine\ORMAdapter;
 use Omines\DataTablesBundle\Column\DateTimeColumn;
 use Omines\DataTablesBundle\Column\NumberColumn;
@@ -12,18 +14,29 @@ use Omines\DataTablesBundle\DataTableFactory;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * @Route("/ticket/statut")
+ * @Route("/customers/{global}/ticket/statut", requirements={"global":"\d+"})
  */
 class TicketStatutController extends AbstractController
 {
+    private $global;
+    private $session;
+
+    public function __construct(SessionInterface $session)
+    {
+        $this->session = $session;
+    }
     /**
      * @Route("/", name="ticket_statut_index", methods={"GET", "POST"})
      */
-    public function index(Request $request, DataTableFactory $dataTableFactory): Response
+    public function index(Request $request, GlobalStatut $global, DataTableFactory $dataTableFactory): Response
     {
+
+        $this->global = $global;
+
         $table = $dataTableFactory->create()
         ->add('id', NumberColumn::class, ['label' => '#'])
         ->add('name', TextColumn::class, ['label' => 'Nom'])
@@ -32,13 +45,30 @@ class TicketStatutController extends AbstractController
                 return $context->getId();
             }, 
             'render' => function($value, $context){
-                $show = sprintf('<a href="%s" class="btn btn-primary">Regarder</a>', $this->generateUrl('ticket_statut_show', ['id' => $value]));
-                $edit = sprintf('<a href="%s" class="btn btn-primary">Modifier</a>', $this->generateUrl('ticket_statut_edit', ['id' => $value]));
+                $show = sprintf('<a href="%s" class="btn btn-primary">Regarder</a>', $this->generateUrl('ticket_statut_show', 
+                [
+                    'id' => $value,
+                    'global' => $this->session->get('global')
+                ]
+                ));
+                $edit = sprintf('<a href="%s" class="btn btn-primary">Modifier</a>', $this->generateUrl('ticket_statut_edit', 
+                [
+                    'id' => $value,
+                    'global' => $this->session->get('global')
+                ]
+                ));
                 return $show.$edit;
             }, 
             'label' => 'Actions'
         ])->createAdapter(ORMAdapter::class, [
-            'entity' => TicketStatut::class
+            'entity' => TicketStatut::class,
+            'query' => function(QueryBuilder $builder) {
+                return $builder
+                ->select('t')
+                ->where('t.global_statut = :g')
+                ->setParameter('g', $this->global)
+                ->from(TicketStatut::class, 't');
+            }
         ])->handleRequest($request);
 
         if ($table->isCallback()) {
@@ -52,18 +82,21 @@ class TicketStatutController extends AbstractController
     /**
      * @Route("/new", name="ticket_statut_new", methods={"GET","POST"})
      */
-    public function new(Request $request): Response
+    public function new(Request $request, GlobalStatut $global): Response
     {
         $ticketStatut = new TicketStatut();
         $form = $this->createForm(TicketStatutType::class, $ticketStatut);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $ticketStatut->setGlobalStatut($global);
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($ticketStatut);
             $entityManager->flush();
 
-            return $this->redirectToRoute('ticket_statut_index');
+            return $this->redirectToRoute('ticket_statut_index', [
+                'global' => $this->session->get('global')
+            ]);
         }
 
         return $this->render('ticket_statut/new.html.twig', [
@@ -75,44 +108,67 @@ class TicketStatutController extends AbstractController
     /**
      * @Route("/{id}", name="ticket_statut_show", methods={"GET"})
      */
-    public function show(TicketStatut $ticketStatut): Response
+    public function show(GlobalStatut $global, TicketStatut $ticketStatut): Response
     {
-        return $this->render('ticket_statut/show.html.twig', [
-            'ticket_statut' => $ticketStatut,
+        if($ticketStatut->getGlobalStatut() === $global){
+            return $this->render('ticket_statut/show.html.twig', [
+                'ticket_statut' => $ticketStatut,
+            ]);
+        }
+        $this->addFlash('error', 'Le statut ticket '.$ticketStatut->getId().' n\'a pas pour statut global '.$global->getId());
+        return $this->redirectToRoute('ticket_statut_index', [
+            'global' => $this->session->get('global')
         ]);
     }
 
     /**
      * @Route("/{id}/edit", name="ticket_statut_edit", methods={"GET","POST"})
      */
-    public function edit(Request $request, TicketStatut $ticketStatut): Response
+    public function edit(Request $request, GlobalStatut $global, TicketStatut $ticketStatut): Response
     {
-        $form = $this->createForm(TicketStatutType::class, $ticketStatut);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
-
-            return $this->redirectToRoute('ticket_statut_index');
+        if($ticketStatut->getGlobalStatut() === $global){
+            $form = $this->createForm(TicketStatutType::class, $ticketStatut);
+            $form->handleRequest($request);
+    
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
+    
+                return $this->redirectToRoute('ticket_statut_index', [
+                    'global' => $this->session->get('global')
+                ]);
+            }
+    
+            return $this->render('ticket_statut/edit.html.twig', [
+                'ticket_statut' => $ticketStatut,
+                'form' => $form->createView(),
+            ]);
         }
 
-        return $this->render('ticket_statut/edit.html.twig', [
-            'ticket_statut' => $ticketStatut,
-            'form' => $form->createView(),
+        $this->addFlash('error', 'Le statut ticket '.$ticketStatut->getId().' n\'a pas pour statut global '.$global->getId());
+        return $this->redirectToRoute('ticket_statut_index', [
+            'global' => $this->session->get('global')
         ]);
     }
 
     /**
      * @Route("/{id}", name="ticket_statut_delete", methods={"DELETE"})
      */
-    public function delete(Request $request, TicketStatut $ticketStatut): Response
+    public function delete(Request $request, GlobalStatut $global, TicketStatut $ticketStatut): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$ticketStatut->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($ticketStatut);
-            $entityManager->flush();
+        if($ticketStatut->getGlobalStatut() === $global){
+            if ($this->isCsrfTokenValid('delete'.$ticketStatut->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($ticketStatut);
+                $entityManager->flush();
+            }
+    
+            return $this->redirectToRoute('ticket_statut_index');
         }
 
-        return $this->redirectToRoute('ticket_statut_index');
+        $this->addFlash('error', 'Le statut ticket '.$ticketStatut->getId().' n\'a pas pour statut global '.$global->getId());
+        return $this->redirectToRoute('ticket_statut_index', [
+            'global' => $this->session->get('global')
+        ]);
+       
     }
 }
